+++ pkg/gui/global_handlers.go
@@ -192,7 +192,7 @@ func (gui *Gui) fetch() (err error) {
 	gui.Mutexes.FetchMutex.Lock()
 	defer gui.Mutexes.FetchMutex.Unlock()
 
-	err = gui.Git.WithSpan("Fetch").Fetch(commands.FetchOptions{})
+	err = gui.Git.WithSpan("Fetch").Sync().Fetch(commands.FetchOptions{})
 
 	if err != nil && strings.Contains(err.Error(), "exit status 128") {
 		_ = gui.CreateErrorPanel(gui.Tr.PassUnameWrong)
@@ -207,7 +207,7 @@ func (gui *Gui) fetchInBackground() (err error) {
 	gui.Mutexes.FetchMutex.Lock()
 	defer gui.Mutexes.FetchMutex.Unlock()
 
-	_ = gui.Git.FetchInBackground(commands.FetchOptions{})
+	_ = gui.Git.Sync().FetchInBackground(commands.FetchOptions{})
 
 	gui.refreshAfterFetch()
